Add actions creator for ingeridients
